refactor: replace jsdoc typedefs with TypeScript types, use folder structure
